refactor: TableGroup에서 List<OrderTable> 필드 제거
